v2v: Add real *.inf files from virtio-win 1.7.4.
These drivers are distributed under GPLv2, which is a freely
redistributable license, and in any case we're not distributing the
code, just the *.inf metadata.

This is important so that we can test parsing real *.inf files, and
also because we need to look at more sections in the file that I
thought previously.

This updates commit c5237e28c5c82a27ea587540baa9c6d58411c4fb.