inspection: Add rules compiler to the generator.
This adds a "prolog-inspired" compiler to the generator, which we can
use to generate guestfs inspection.  The compiler generates C code
which should be reasonably efficient.

Note this change requires ocamllex and ocamlyacc, but these are
normally part of any OCaml distribution, since they are part of the
OCaml compiler.  We also import some GPL-licensed gnulib code for
handling sets, which is OK because we're going to use these in a
standalone program (in following commits).