appliance: Pass root=UUID=... to supermin.
Instead of mounting the appliance by device name (eg. root=/dev/sdb),
mount the appliance filesystem by filesystem UUID.  This allows the
appliance to be robust against the non-determinism of SCSI device
enumeration, but it does require a corresponding change to be added to
supermin, and therefore a new version of supermin.

I did not make a corresponding change to UML as it is unlikely to have
a problem with enumerating ubd devices.