NOT COMPLETED Reimplement disk labels.
Libguestfs allows you to associate a disk label with a drive when
adding the drive.  This makes the disk appear under the device name
/dev/disk/guestfs/<LABEL> as well as the normal device name like
/dev/sdX.  Also partitions appear as /dev/disk/guestfs/<LABEL>[1..].
Disk labels are optional for ordinary disks, but required for
hot-plugged disks.

Previously we implemented disk labels using the qemu disk serial
number and some udev rules.

This commit reimplements disk labels using some internal APIs:

  guestfs_internal_set_disk_label (g, index, label)
    - called at launch for static disks, and when hot-adding a disk

  guestfs_internal_clear_disk_label (g, index)
    - called when hot-removing a disk

udev rules are no longer used, so the actual /dev/disk/guestfs paths
are not created.  Instead some magic is done during device name
translation to translate the special paths into the correct device or
partition name.

There are a couple of reasons for this change:

 - Allow serial numbers to be used for stable device naming of
   virtio-scsi devices (in a future commit).

 - Allow disk labels to be used in places where serial numbers are not
   supported, ie. UML.