lib: drives: Pass the drive name as the SCSI serial number.
Pass the drive name to the appliance via the SCSI serial number.  For
example the API drive called "/dev/sda" gets the scsi serial number
"GUESTFS_a" and that in turn is mapped to a particular /dev/disk/by-id
symlink in the appliance, eg:

$ ls -l /dev/disk/by-id
[...]
lrwxrwxrwx 1 root root 11 Apr 25 15:08 scsi-0QEMU_QEMU_HARDDISK_GUESTFS_a -> ../../sdc

Because SCSI device enumeration is not deterministic, the drive in the
appliance may not be the same (as shown above).