daemon: Add guestfs-inspection program.
This implements a program which performs guestfs inspection.  The
program runs inside the appliance, so it is faster than doing
inspection from the library side through the API.  This commit does
not enable the new inspection code (see following commits for that).

As a side effect of this implementation, some pieces of daemon code
are reimplemented in OCaml, as well as the library side
‘guestfs_list_filesystems’ API.  In this commit they do not replace
those functions in the daemon (but see following commits).  The
replacements share the same filename, eg:

  blkid.c  == is partially replaced by ==>   blkid.ml, blkid.mli
  lvm2.c   =======>   lvm2.ml, lvm2.mli

Note that the OCaml compiler (either ocamlc or ocamlopt) is now
required even for building from tarballs.