daemon: Allow parts of the daemon to be written in OCaml.
This change allows parts of the daemon to be written in the OCaml
programming language.  I am using the ‘Main Program in C’ method along
with ‘-output-obj’ to create an object file from the OCaml code /
runtime, as described here:
https://caml.inria.fr/pub/docs/manual-ocaml/intfc.html

This change doesn't make any functional changes yet.  There is simply
an empty ‘daemon.ml’ file.

Note that the OCaml compiler (either ocamlc or ocamlopt) is now
required even for building from tarballs.