build: Don't test for qemu virtio-serial at configure time.
qemu has supported virtio-serial since forever, and we depend on much
newer features now, so testing for virtio-serial is pointless.

In fact allowing a set of possible qemu binaries to be supplied is
also pointless, since packagers or even people building from source
know what qemu binary they want to use.

Now the possible options for --with-qemu are:

* --with-qemu=search (or --with-qemu omitted)
    Search for various qemu binaries on the path.

* --with-qemu=/path/to/qemu
    Use the supplied qemu binary without question.  No checks on
    the binary are done.

* --with-qemu=no or --without-qemu
    Disable the default hypervisor.  You have to set $LIBGUESTFS_HV
    before running programs.