inspection: Add network interfaces to inspection data.
This adds network interfaces to inspection data.

The data is returned as a flat list of structs containing a general
overview of the interfaces, and a separate call returning a fairly
arbitrary list of key/value parameters for each interface.  The list
of parameters is roughly modelled on the ifcfg-* files, although with
some mapping of the data to make it easier to consume.

Note this doesn't attempt a detailed model of all possible guest
network configurations, which in particular cases could be extremely
complex.

This commits just throws ENOTSUP error for all guests.  Following
commits add this information for Red Hat-derived, Debian-derived and
Windows guests.