daemon: Allow OCaml code to raise Not_supported exception.
It would be possible to implement this as:

  raise (Unix_error (ENOTSUP, fn, arg))

except that Unix.ENOTSUP is not defined (OCaml 4.07).  Instead of
working around this let's have a special exception, and anyway this
works like the C code.

The only tricky part of this patch is that I wanted to use the
‘Daemon’ module name for daemonic things, such as this exception.  So
I had to rename the existing Daemon module (which contains
initialization code) to ‘Daemon_init’.